Refactors navbar assets files
