+++ src/app/top/navbar/Navbar.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react'
 import { NavLink, useLocation } from 'react-router-dom'
-import { IconSmallPlus } from '../global/icons/IconSmallPlus'
+import { IconSmallPlus } from '../../global/icons/IconSmallPlus'
 import './style/social-media'
 import './style/navbar'
